linera_bridge::evm

Module light_client

LightClient Solidity ABI (committee/epoch management on EVM).

Structs

addCommitteeCall

Function with signature addCommittee(bytes,bytes,bytes[]) and selector 0xacff1caf.

addCommitteeReturn

Container type for the return parameters of the addCommittee(bytes,bytes,bytes[]) function.

currentEpochCall

Function with signature currentEpoch() and selector 0x76671808.

currentEpochReturn

Container type for the return parameters of the currentEpoch() function.

verifyBlockCall

Function with signature verifyBlock(bytes) and selector 0x66f99037.

verifyBlockReturn

Container type for the return parameters of the verifyBlock(bytes) function.

Constants

SOURCE

Solidity source for the LightClient contract.