linera_bridge::evm

Module microchain

Microchain Solidity ABI (block relay on EVM).

Structs

addBlockCall

Function with signature addBlock(bytes) and selector 0x054f12ad.

addBlockReturn

Container type for the return parameters of the addBlock(bytes) function.

chainIdCall

Function with signature chainId() and selector 0x9a8a0592.

chainIdReturn

Container type for the return parameters of the chainId() function.

lightClientCall

Function with signature lightClient() and selector 0xb5700e68.

lightClientReturn

Container type for the return parameters of the lightClient() function.

nextExpectedHeightCall

Function with signature nextExpectedHeight() and selector 0xbd6894f8.

nextExpectedHeightReturn

Container type for the return parameters of the nextExpectedHeight() function.

Constants

SOURCE

Solidity source for the Microchain abstract contract.